extract steps
